+++ service/hibernate/ServiceTypeDaoHibernate.java
@@ -3,7 +3,6 @@ package wbs.platform.service.hibernate;
 import java.util.List;
 
 import lombok.NonNull;
-
 import wbs.framework.hibernate.HibernateDao;
 import wbs.platform.service.model.ServiceTypeDao;
 import wbs.platform.service.model.ServiceTypeRec;
@@ -35,7 +34,7 @@ class ServiceTypeDaoHibernate
 
 		return get (
 			ServiceTypeRec.class,
-			(int) (long) serviceTypeId);
+			serviceTypeId);
 
 	}
